+++ qa/software/API_Test_Suite/src/caarray/client/test/suite/ArrayDesignTestSuite.java
@@ -155,7 +155,7 @@ public class ArrayDesignTestSuite extends SearchByExampleTestSuite
         search.setArrayDesign(example);
         if (headerIndexMap.get(TEST_CASE) < input.length
                 && !input[headerIndexMap.get(TEST_CASE)].equals(""))
-            search.setTestCase(Integer.parseInt(input[headerIndexMap.get(TEST_CASE)]
+            search.setTestCase(Float.parseFloat(input[headerIndexMap.get(TEST_CASE)]
                     .trim()));
         if (headerIndexMap.get(EXPECTED_RESULTS) < input.length
                 && !input[headerIndexMap.get(EXPECTED_RESULTS)].equals(""))
