+++ qa/software/API_Test_Suite/src/caarray/client/test/suite/AssayTypeTestSuite.java
@@ -59,7 +59,7 @@ public class AssayTypeTestSuite extends SearchByExampleTestSuite
         search.setAssayType(example);
         if (headerIndexMap.get(TEST_CASE) < input.length
                 && !input[headerIndexMap.get(TEST_CASE)].equals(""))
-            search.setTestCase(Integer.parseInt(input[headerIndexMap.get(TEST_CASE)]
+            search.setTestCase(Float.parseFloat(input[headerIndexMap.get(TEST_CASE)]
                     .trim()));
         if (headerIndexMap.get(EXPECTED_RESULTS) < input.length
                 && !input[headerIndexMap.get(EXPECTED_RESULTS)].equals(""))
