+++ qa/software/API_Test_Suite/src/caarray/client/test/suite/OrganismTestSuite.java
@@ -218,7 +218,7 @@ public class OrganismTestSuite extends SearchByExampleTestSuite
         search.setOrganism(example);
         if (headerIndexMap.get(TEST_CASE) < input.length
                 && !input[headerIndexMap.get(TEST_CASE)].equals(""))
-            search.setTestCase(Integer.parseInt(input[headerIndexMap
+            search.setTestCase(Float.parseFloat(input[headerIndexMap
                     .get(TEST_CASE)].trim()));
         if (headerIndexMap.get(EXPECTED_RESULTS) < input.length
                 && !input[headerIndexMap.get(EXPECTED_RESULTS)].equals(""))
